Don't close stream in bus gc finalizer
This means that a bus object going out of scope will unfortunately not
automatically clean up the async listener, and cause a script to hang
when the top-level finishes evaluation.

The issue is that a top-level error will trigger a hard teardown via
janet_clear_memory, which iterates through the allocated memory blocks
calling janet_deinit_block. This leaves a bunch of dangling
pointers, so we can't safely interact with other Janet objects inside
the bus finalizer. For example, fiber->ev_state may have already been
free'd which will crash when we try to call janet_async_end.